+++ packages/components/CHANGELOG.md
@@ -1,4 +1,4 @@
-### 0.5.1 - 0.5.10
+### 0.5.1 - 0.5.8
 fix bugs for new project structure
 * Avatar: fix a issue of group that the margin-left doesn't take effect
 * DatePicker: remove the store implementation
@@ -11,6 +11,7 @@ fix bugs for new project structure
 * Toggle: remove box-shadow property
 * Grid Col: support flex 
 * IconInput: fix disabled issue
+* TimePicker: finished
 
 ### 0.5.0
 * Toggle: support form's defaultValue, other components need to double check as well